Move `configtest` command to `test config`
