+++ libbeat/cmd/root.go
@@ -27,7 +27,6 @@ type BeatsRootCmd struct {
 	RunCmd        *cobra.Command
 	SetupCmd      *cobra.Command
 	VersionCmd    *cobra.Command
-	ConfigTestCmd *cobra.Command
 	CompletionCmd *cobra.Command
 	ExportCmd     *cobra.Command
 	TestCmd       *cobra.Command
@@ -51,7 +50,6 @@ func GenRootCmdWithRunFlags(name, version string, beatCreator beat.Creator, runF
 	rootCmd.RunCmd = genRunCmd(name, version, beatCreator, runFlags)
 	rootCmd.SetupCmd = genSetupCmd(name, version, beatCreator)
 	rootCmd.VersionCmd = genVersionCmd(name, version)
-	rootCmd.ConfigTestCmd = genConfigTestCmd(name, version, beatCreator)
 	rootCmd.CompletionCmd = genCompletionCmd(name, version, rootCmd)
 	rootCmd.ExportCmd = genExportCmd(name, version, beatCreator)
 	rootCmd.TestCmd = genTestCmd(name, version, beatCreator)
@@ -79,7 +77,6 @@ func GenRootCmdWithRunFlags(name, version string, beatCreator beat.Creator, runF
 	rootCmd.AddCommand(rootCmd.RunCmd)
 	rootCmd.AddCommand(rootCmd.SetupCmd)
 	rootCmd.AddCommand(rootCmd.VersionCmd)
-	rootCmd.AddCommand(rootCmd.ConfigTestCmd)
 	rootCmd.AddCommand(rootCmd.CompletionCmd)
 	rootCmd.AddCommand(rootCmd.ExportCmd)
 	rootCmd.AddCommand(rootCmd.TestCmd)
